requirements.txt
tweepy==2.3.0
requests==2.3.0
pillow==2.8.1
